+++ CHANGELOG.md
@@ -6,3 +6,9 @@ Toutes les modifications notables seront listées ici.
 - Version initiale de CodeScribe.
 - Export des projets C#, Angular, Python en Markdown.
 - Options `--export-txt`, `--ignore-spec`, `--minimal`, `--txt`
+
+## [1.1.0] - 2025-05-28
+- Ajout de l'option `--exclude-ext` pour exclure des extensions.
+- Ajout des options `--version` et `--default-ext`.
+- Ajout de l'option `--git-ignore` pour exclure selon le fichier `.gitignore`.
+- Ajout de l'option `--exclude-dir` pour ignorer des répertoires.